fix remaining nightly tests
